refactor: Move config model to customer analytics product app
