+++ products/customer_analytics/backend/models/__init__.py
@@ -1,0 +1,5 @@
+from .team_customer_analytics_config import TeamCustomerAnalyticsConfig
+
+__all__ = [
+    "TeamCustomerAnalyticsConfig",
+]
