+++ products/customer_analytics/backend/models/team_customer_analytics_config.py
